requirements.txt
Flask==2.0.1
   Flask-SQLAlchemy==2.5.1
   Flask-WTF==0.15.1
   discord.py==2.0.0
   python-dotenv==0.19.0
   tradingview-ta==3.2.10
   requests==2.26.0
   selenium==3.141.0
   aiohttp==3.7.4
   gunicorn==23.0.0